Extract DOM processing from XML config classes
This change extracts the DOM processing parts from the XML config
classes in order to make it reuseable for YAML configuration.

The change consists only moving methods.

The driving changes are:
- ClientXmlElements renamed to ClientConfigSections
- XmlElements renamed to ConfigSections
- Processing client config DOM nodes moved from XmlClientConfigBuilder
to ClientDomConfigProcessor
- Processing member config DOM nodes moved from XmlConfigBuilder
to MemberDomConfigProcessor
- *DomConfigProcessors are used by *ConfigBuilders instead of being
extended by them
- DOM utility methods are moved from the AbstractXmlConfigHelper root
class to DomConfigHelper utility class
