+++ hazelcast-spring/src/main/java/com/hazelcast/spring/CacheManagerBeanDefinitionParser.java
@@ -26,6 +26,9 @@ import org.w3c.dom.Node;
 
 import java.util.Properties;
 
+import static com.hazelcast.config.DomConfigHelper.childElements;
+import static com.hazelcast.config.DomConfigHelper.cleanNodeName;
+
 /**
  * Bean definition parser for JCache {@link javax.cache.CacheManager}.
  * <p>
