+++ hazelcast/src/test/java/com/hazelcast/config/IterableNodeListTest.java
@@ -33,8 +33,8 @@ import javax.xml.parsers.ParserConfigurationException;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 
-import static com.hazelcast.config.AbstractXmlConfigHelper.asElementIterable;
-import static com.hazelcast.config.AbstractXmlConfigHelper.cleanNodeName;
+import static com.hazelcast.config.DomConfigHelper.asElementIterable;
+import static com.hazelcast.config.DomConfigHelper.cleanNodeName;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
