+++ hazelcast/src/test/java/com/hazelcast/config/MyServiceConfigParser.java
@@ -20,6 +20,10 @@ import com.hazelcast.spi.ServiceConfigurationParser;
 import org.w3c.dom.Element;
 import org.w3c.dom.Node;
 
+import static com.hazelcast.config.DomConfigHelper.childElements;
+import static com.hazelcast.config.DomConfigHelper.cleanNodeName;
+import static org.ops4j.util.xml.XmlUtils.getTextContent;
+
 public class MyServiceConfigParser extends AbstractXmlConfigHelper implements ServiceConfigurationParser<MyServiceConfig> {
 
     public MyServiceConfig parse(Element element) {
